move code related to histogram definitions to a separate file
